unblocked shit for people in school :)
I will link alot of unblocked websites game etc etc. This is handy for people on school like me so the unblocked stuff for my school. My school block their shit on the internet not on the machine itself. So I may do not add stuff that is unblocked for you but for me it is blocked :(. I hope this will help I will make the thing this week and then you guys have fun!!